Replace ${NAME} with plain `mylib`.
There aren't good reasons for doing this except to accommodate a
workflow of copy-pasting CMakeLists.txt around different projects.
But this is fickle and will accumulate and propagate bugs over
time.

Besides, it comes with a lot of downsides. It makes the code less
searchable, it doesn't scale to multiple targets, and it generally
makes the code harder to read (especially in the presence of multiple
targets).

Basically, if you want to reuse code, write a function().